refactor(templating): extract maskOrDecryptVaultDataIfNecessary into its own file
