+++ packages/insomnia/src/common/render.ts
@@ -17,6 +17,7 @@ import orderedJSON from 'json-order';
 import { getOrInheritAuthentication, getOrInheritHeaders } from '../network/network';
 import { getRuntime } from '../runtimes';
 import { NUNJUCKS_TEMPLATE_GLOBAL_PROPERTY_NAME } from '../templating/constants';
+import { maskOrDecryptVaultDataIfNecessary } from '../templating/mask-or-decrypt-vault-data';
 import { RenderError } from '../templating/render-error';
 import type {
   BaseRenderContext,
@@ -26,7 +27,6 @@ import type {
   RenderedRequest,
 } from '../templating/types';
 import * as templatingUtils from '../templating/utils';
-import { maskOrDecryptVaultDataIfNecessary } from '../templating/utils';
 import { setDefaultProtocol } from '../utils/url/protocol';
 import { CONTENT_TYPE_GRAPHQL, JSON_ORDER_SEPARATOR } from './constants';
 import { database as db } from './database';
